rename motor hardware files
