+++ gripper/firmware/interfaces_grip_motor.cpp
@@ -139,9 +139,7 @@ extern "C" void gripper_enc_idle_state_callback_glue(bool val) {
 }
 
 void grip_motor_iface::initialize() {
-    // Initialize DAC
-    initialize_dac();
-    initialize_enc();
+    initialize_hardware_g();
     set_brushed_motor_timer_callback(call_brushed_motor_handler,
                                      gripper_enc_overflow_callback_glue,
                                      gripper_enc_idle_state_callback_glue);
