--- gripper/firmware/motor_hardware.c
@@ -1,143 +1,0 @@
-#include "motor_hardware.h"
-
-DAC_HandleTypeDef hdac1;
-
-
-void HAL_SPI_MspInit(SPI_HandleTypeDef* hspi) {
-    GPIO_InitTypeDef GPIO_InitStruct = {0};
-    if (hspi->Instance == SPI2) {
-        /* Peripheral clock enable */
-        __HAL_RCC_SPI2_CLK_ENABLE();
-        __HAL_RCC_GPIOA_CLK_ENABLE();
-        __HAL_RCC_GPIOB_CLK_ENABLE();
-        __HAL_RCC_GPIOC_CLK_ENABLE();
-        /**SPI2 GPIO Configuration
-        PB12  ------> SPI2_CS
-        PB13  ------> SPI2_SCK
-        PB14  ------> SPI2_MISO
-        PB15  ------> SPI2_MOSI
-        Step/Dir
-        PB10   ------> Motor Dir Pin
-        PB1   ------> Motor Step Pin
-        Enable
-        PA9   ------> Motor Enable Pin
-        */
-        GPIO_InitStruct.Pin = GPIO_PIN_13 | GPIO_PIN_14 | GPIO_PIN_15;
-        GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-        GPIO_InitStruct.Alternate = GPIO_AF5_SPI2;
-        HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-        // Chip select
-        GPIO_InitStruct.Pin = GPIO_PIN_12;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-        // Dir
-        GPIO_InitStruct.Pin = GPIO_PIN_10;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-        HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-        // Step
-        GPIO_InitStruct.Pin = GPIO_PIN_1;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        HAL_GPIO_Init(GPIOB,  // NOLINT(cppcoreguidelines-pro-type-cstyle-cast)
-                      &GPIO_InitStruct);
-
-        // Enable
-        GPIO_InitStruct.Pin = GPIO_PIN_9;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        HAL_GPIO_Init(GPIOA,  // NOLINT(cppcoreguidelines-pro-type-cstyle-cast)
-                      &GPIO_InitStruct);
-    }
-}
-SPI_HandleTypeDef hspi2 = {
-    .Instance = SPI2,
-    .Init = {.Mode = SPI_MODE_MASTER,
-             .Direction = SPI_DIRECTION_2LINES,
-             .DataSize = SPI_DATASIZE_8BIT,
-             .CLKPolarity = SPI_POLARITY_HIGH,
-             .CLKPhase = SPI_PHASE_2EDGE,
-             .NSS = SPI_NSS_SOFT,
-             .BaudRatePrescaler = SPI_BAUDRATEPRESCALER_32,
-             .FirstBit = SPI_FIRSTBIT_MSB,
-             .TIMode = SPI_TIMODE_DISABLE,
-             .CRCCalculation = SPI_CRCCALCULATION_DISABLE,
-             .CRCPolynomial = 7,
-             .CRCLength = SPI_CRC_LENGTH_DATASIZE,
-             .NSSPMode = SPI_NSS_PULSE_DISABLE}
-
-};
-
-/**
- * @brief SPI MSP De-Initialization
- * This function freeze the hardware resources used in this example
- * @param hspi: SPI handle pointer
- * @retval None
- */
-void HAL_SPI_MspDeInit(SPI_HandleTypeDef* hspi) {
-    if (hspi->Instance == SPI2) {
-        /* Peripheral clock disable */
-        __HAL_RCC_SPI2_CLK_DISABLE();
-        HAL_GPIO_DeInit(GPIOB, GPIO_PIN_12 | GPIO_PIN_13 | GPIO_PIN_14 |
-                                   GPIO_PIN_15 | GPIO_PIN_1 | GPIO_PIN_10);
-        HAL_GPIO_DeInit(GPIOA, GPIO_PIN_9);
-    }
-}
-
-HAL_StatusTypeDef initialize_spi() {
-    __HAL_RCC_SPI2_CLK_ENABLE();
-    return HAL_SPI_Init(&hspi2);
-}
-
-/**
- * @brief DAC1 Initialization Function
- * @param None
- * @retval None
- */
-static void MX_DAC1_Init(void) {
-    hdac1.Instance = DAC1;
-    if (HAL_DAC_Init(&hdac1) != HAL_OK) {
-        Error_Handler();
-    }
-}
-
-/**
- * @brief DAC MSP Initialization
- * This function configures the hardware resources used in this example
- * @param hdac: DAC handle pointer
- * @retval None
- */
-void HAL_DAC_MspInit(DAC_HandleTypeDef* hdac) {
-    GPIO_InitTypeDef GPIO_InitStruct = {0};
-    if (hdac->Instance == DAC1) {
-        /* Peripheral clock enable */
-        __HAL_RCC_DAC1_CLK_ENABLE();
-        __HAL_RCC_GPIOA_CLK_ENABLE();
-        /**DAC1 GPIO Configuration
-        PA4     ------> DAC1_OUT1
-        */
-        GPIO_InitStruct.Pin = GPIO_PIN_4;
-        GPIO_InitStruct.Mode = GPIO_MODE_ANALOG;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
-    }
-}
-
-/**
- * @brief DAC MSP De-Initialization
- * This function freeze the hardware resources used in this example
- * @param hdac: DAC handle pointer
- * @retval None
- */
-void HAL_DAC_MspDeInit(DAC_HandleTypeDef* hdac) {
-    if (hdac->Instance == DAC1) {
-        __HAL_RCC_DAC1_CLK_DISABLE();
-        HAL_GPIO_DeInit(GPIOA, GPIO_PIN_4);
-    }
-}
-
-void initialize_dac() { MX_DAC1_Init();}
